+++ src/App.css
@@ -4,7 +4,6 @@ body, html {
 }
 .App {
   text-align: center;
-  border: 1px solid red;
   min-height: 100vh;
   display: flex;
   justify-content: center;
